First 4 Hashes Correct
Broke out chunk write data to combinational logic.
Changed update condition to always during CHUNK_FILL (waiting for
mem_data not an issue, as it won't advance during the wait)
Changed ca_state to evaluate when the next state is CHUNK_READY, also to
jump to END_WORD when the current word finishes the data.